refactor: `U32Word` to a more abstract purpose impl

+++ wks/hash_functions/rs_sha1/src/sha1hasher.rs
@@ -59,7 +59,7 @@ impl Hasher for Sha1Hasher {
             bytes = &bytes[left as usize..];
         }
 
-        while bytes.len() >= Self::U8_PADDING_COUNT as usize {
+        while bytes.len() >= Self::U8_PADDING_COUNT {
             self.padding.clone_from_slice(&bytes[..Self::U8_PADDING_COUNT]);
             Self::hash_block(&self.padding, &mut self.state);
             bytes = &bytes[Self::U8_PADDING_COUNT..];
@@ -76,7 +76,7 @@ impl HasherContext for Sha1Hasher {
 
     fn finish(&mut self) -> Self::State {
         let zero_padding_length = Self::zeros_pad_length(self.size as usize);
-        let mut offset_pad: [u8; Self::U8_PADDING_COUNT as usize] = [0u8; Self::U8_PADDING_COUNT as usize];
+        let mut offset_pad: [u8; Self::U8_PADDING_COUNT] = [0u8; Self::U8_PADDING_COUNT];
         offset_pad[0] = 0x80;
 
         let len = self.size;
